Make library scanning method blocking
And a shitload of other changes to accommodate this. They should have
been in different commits but the change is to big and interwined by
now.
